refactor: about 에서 makers로 이름 일괄 변경
