--- components/about/data/index.ts
@@ -1,4 +1,0 @@
-import { generation1 } from '@/components/about/data/generation1';
-import { MakersGeneration } from '@/components/about/data/types';
-
-export const makersGenerationsData: MakersGeneration[] = [generation1];
