+++ components/makers/TeamBlock.stories.tsx
@@ -1,7 +1,7 @@
 import { ComponentMeta, ComponentStory } from '@storybook/react';
 
-import { WithImage } from '@/components/about/PersonBlock.stories';
-import TeamBlock from '@/components/about/TeamBlock';
+import { WithImage } from '@/components/makers/PersonBlock.stories';
+import TeamBlock from '@/components/makers/TeamBlock';
 
 export default {
   component: TeamBlock,
