+++ components/makers/data/index.ts
@@ -1,0 +1,4 @@
+import { generation1 } from '@/components/makers/data/generation1';
+import { MakersGeneration } from '@/components/makers/data/types';
+
+export const makersGenerationsData: MakersGeneration[] = [generation1];
